Fixed some of the qbwriter unit tests after the refactor.
