+++ tests/unit/writers/qbwriter/test_qbwriter.py
@@ -1,6 +1,4 @@
 from tempfile import TemporaryDirectory
-from typing import List
-from copy import deepcopy
 import csv
 from pathlib import Path
 
@@ -45,480 +43,8 @@ class TestQbMeasure(QbMeasure, UriIdentifiable):
         pass
 
 
-def _get_standard_cube_for_columns(columns: List[CsvColumn]) -> Cube:
-    data = pd.DataFrame(
-        {
-            "Country": ["Wales", "Scotland", "England", "Northern Ireland"],
-            "Observed Value": [101.5, 56.2, 12.4, 77.8],
-            "Marker": ["Provisional", "Provisional", "Provisional", "Provisional"],
-        }
-    )
-    metadata: CatalogMetadata = CatalogMetadata("Cube Name")
-
-    return Cube(deepcopy(metadata), data.copy(deep=True), columns)
-
-
-def _assert_component_defined(
-    dataset: rdf.qb.DataSet, name: str
-) -> rdf.qb.ComponentSpecification:
-    component = first(
-        dataset.structure.components,
-        lambda x: str(x.uri) == f"cube-name.csv#component/{name}",
-    )
-    assert component is not None
-    return component
-
-
-def _assert_component_property_defined(
-    component: rdf.qb.ComponentSpecification, property_uri: str
-) -> None:
-    property = first(
-        component.componentProperties, lambda x: str(x.uri) == property_uri
-    )
-    assert property is not None
-    return property
-
-
 empty_cube = Cube(CatalogMetadata("Cube Name"), pd.DataFrame, [])
 empty_qbwriter = QbWriter(empty_cube)
-
-
-def test_structure_defined():
-    cube = _get_standard_cube_for_columns(
-        [
-            QbColumn(
-                "Country",
-                ExistingQbDimension("http://example.org/dimensions/country"),
-            ),
-            QbColumn(
-                "Marker",
-                ExistingQbAttribute("http://example.org/attributes/marker"),
-            ),
-            QbColumn(
-                "Observed Value",
-                QbObservationValue(
-                    ExistingQbMeasure("http://example.org/units/some-existing-measure"),
-                    ExistingQbUnit("http://example.org/units/some-existing-unit"),
-                ),
-            ),
-        ]
-    )
-
-    qbwriter = QbWriter(cube)
-    dataset = qbwriter._generate_qb_dataset_dsd_definitions()
-
-    assert dataset is not None
-
-    assert dataset.structure is not None
-    assert type(dataset.structure) == rdf.qb.DataStructureDefinition
-
-    assert dataset.structure.componentProperties is not None
-
-    _assert_component_defined(dataset, "country")
-    _assert_component_defined(dataset, "marker")
-    _assert_component_defined(dataset, "unit")
-    _assert_component_defined(dataset, "some-existing-measure")
-
-
-def test_structure_uri():
-    cube = Cube(CatalogMetadata("Cube Name"))
-    qbwriter = QbWriter(cube)
-
-    actual = qbwriter._generate_qb_dataset_dsd_definitions().structure.uri_str
-    assert actual.startswith("cube-name.csv#")
-
-
-def test_structure_uri_standard_pattern():
-    cube = Cube(CatalogMetadata("Cube Name"), uri_style=URIStyle.Standard)
-    qbwriter = QbWriter(cube)
-
-    actual = qbwriter._generate_qb_dataset_dsd_definitions().structure.uri_str
-    assert actual.startswith("cube-name.csv#")
-
-
-def test_structure_uri_withoutFileExtensions_pattern():
-    cube = Cube(CatalogMetadata("Cube Name"), uri_style=URIStyle.WithoutFileExtensions)
-    qbwriter = QbWriter(cube)
-
-    actual = qbwriter._generate_qb_dataset_dsd_definitions().structure.uri_str
-    assert actual.startswith("cube-name#")
-
-
-def test_generating_concept_uri_template_from_legacy_global_concept_scheme_uri():
-    """
-    Given a globally defined skos:ConceptScheme's URI, generate the URI template for a column which maps the
-    column's value to a concept defined inside the concept scheme.
-    """
-    column = SuppressedCsvColumn("Some Column")
-    code_list = ExistingQbCodeList(
-        "http://base-uri/concept-scheme/this-concept-scheme-name"
-    )
-
-    actual_concept_template_uri = (
-        empty_qbwriter._get_default_value_uri_for_code_list_concepts(column, code_list)
-    )
-    assert (
-        "http://base-uri/concept-scheme/this-concept-scheme-name/{+some_column}"
-        == actual_concept_template_uri
-    )
-
-
-def test_generating_concept_uri_template_from_legacy_local_concept_scheme_uri():
-    """
-    Given a dataset-local skos:ConceptScheme's URI, generate the URI template for a column which maps the
-    column's value to a concept defined inside the concept scheme.
-    """
-    column = SuppressedCsvColumn("Some Column")
-    code_list = ExistingQbCodeList(
-        "http://base-uri/dataset-name#scheme/that-concept-scheme-name"
-    )
-
-    actual_concept_template_uri = (
-        empty_qbwriter._get_default_value_uri_for_code_list_concepts(column, code_list)
-    )
-    assert (
-        "http://base-uri/dataset-name#concept/that-concept-scheme-name/{+some_column}"
-        == actual_concept_template_uri
-    )
-
-
-def test_generating_concept_uri_template_from_csvcubed_concept_scheme_uri():
-    """
-    Given a csvcubed-style skos:ConceptScheme's URI, generate the URI template for a column which maps the
-    column's value to a concept defined inside the concept scheme.
-    """
-    column = SuppressedCsvColumn("Some Column")
-    code_list = ExistingQbCodeList(
-        f"http://base-uri/concept-scheme#{SCHEMA_URI_IDENTIFIER}"
-    )
-
-    actual_concept_template_uri = (
-        empty_qbwriter._get_default_value_uri_for_code_list_concepts(column, code_list)
-    )
-    assert (
-        "http://base-uri/concept-scheme#{+some_column}" == actual_concept_template_uri
-    )
-
-
-def test_generating_concept_uri_template_from_unexpected_concept_scheme_uri():
-    """
-    Given a skos:ConceptScheme's URI *that does not follow the global or dataset-local conventions* used in our
-    tooling, return the column's value as our best guess at the concept's URI.
-    """
-    column = SuppressedCsvColumn("Some Column")
-    code_list = ExistingQbCodeList(
-        "http://base-uri/dataset-name#codes/that-concept-scheme-name"
-    )
-
-    actual_concept_template_uri = (
-        empty_qbwriter._get_default_value_uri_for_code_list_concepts(column, code_list)
-    )
-    assert "{+some_column}" == actual_concept_template_uri
-
-
-def test_default_property_value_uris_existing_dimension_column():
-    """
-    When an existing dimension is used, we can provide the `propertyUrl`, but we cannot guess the `valueUrl`.
-    """
-    column = QbColumn(
-        "Some Column",
-        ExistingQbDimension("http://base-uri/dimensions/existing-dimension"),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "http://base-uri/dimensions/existing-dimension" == default_property_uri
-    assert "{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_new_dimension_column_without_code_list():
-    """
-    When a new dimension is defined without a code list, we can provide the `propertyUrl`,
-    but we cannot guess the `valueUrl`.
-    """
-    column = QbColumn("Some Column", NewQbDimension("Some New Dimension"))
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "cube-name.csv#dimension/some-new-dimension" == default_property_uri
-    assert "{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_new_dimension_column_with_code_list():
-    """
-    When an new dimension is defined with a code list, we can provide the `propertyUrl` and the `valueUrl`.
-    """
-    column = QbColumn(
-        "Some Column",
-        NewQbDimension(
-            "Some New Dimension",
-            code_list=ExistingQbCodeList("http://base-uri/concept-scheme/this-scheme"),
-        ),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "cube-name.csv#dimension/some-new-dimension" == default_property_uri
-    assert (
-        "http://base-uri/concept-scheme/this-scheme/{+some_column}" == default_value_uri
-    )
-
-
-def test_default_property_value_uris_new_dimension_column_with_new_code_list():
-    """
-    When a new dimension is defined with a new code list, by default it should provide standard-formatted property and value Urls
-    """
-    column = QbColumn(
-        "Some Column",
-        NewQbDimension(
-            "Some New Dimension",
-            code_list=NewQbCodeList(CatalogMetadata("Some Catalog"), []),
-        ),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "cube-name.csv#dimension/some-new-dimension" == default_property_uri
-    assert "some-catalog.csv#{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_new_dimension_column_with_new_code_list_for_cube_WithoutFileExtensions_uri_style():
-    """
-    When a new dimension is defined with a new code list, and the cube has a defined uri_style of WithoutFileExtensions,
-    it should provide property and value Urls which follow the cube's uri style
-    """
-
-    uri_styled_cube = Cube(
-        CatalogMetadata("Cube Name"), pd.DataFrame, [], URIStyle.WithoutFileExtensions
-    )
-    uri_styled_qbwriter = QbWriter(uri_styled_cube)
-
-    column = QbColumn(
-        "Some Column",
-        NewQbDimension(
-            "Some New Dimension",
-            code_list=NewQbCodeList(CatalogMetadata("Some Catalog"), []),
-        ),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = uri_styled_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "cube-name#dimension/some-new-dimension" == default_property_uri
-    assert "some-catalog#{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_existing_attribute_existing_values():
-    """
-    When an existing attribute is used, we can provide the `propertyUrl`, but we cannot guess the `valueUrl`.
-    """
-    column = QbColumn(
-        "Some Column",
-        ExistingQbAttribute("http://base-uri/attributes/existing-attribute"),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "http://base-uri/attributes/existing-attribute" == default_property_uri
-    assert "{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_existing_attribute_new_values():
-    """
-    When an existing attribute is used, we can provide the `propertyUrl`, but we cannot guess the `valueUrl`.
-    """
-    column = QbColumn(
-        "Some Column",
-        ExistingQbAttribute(
-            "http://base-uri/attributes/existing-attribute",
-            new_attribute_values=[NewQbAttributeValue("Some Attribute Value")],
-        ),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "http://base-uri/attributes/existing-attribute" == default_property_uri
-    assert "cube-name.csv#attribute/some-column/{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_new_attribute_existing_values():
-    """
-    When a new attribute is defined, we can provide the `propertyUrl`, but we cannot guess the `valueUrl`.
-    """
-    column = QbColumn("Some Column", NewQbAttribute("This New Attribute"))
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "cube-name.csv#attribute/this-new-attribute" == default_property_uri
-    assert "{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_new_attribute_new_values():
-    """
-    When a new attribute is defined, we can provide the `propertyUrl`, but we cannot guess the `valueUrl`.
-    """
-    column = QbColumn(
-        "Some Column",
-        NewQbAttribute(
-            "This New Attribute",
-            new_attribute_values=[NewQbAttributeValue("Something")],
-        ),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "cube-name.csv#attribute/this-new-attribute" == default_property_uri
-    assert (
-        "cube-name.csv#attribute/this-new-attribute/{+some_column}" == default_value_uri
-    )
-
-
-def test_default_property_value_uris_multi_units_all_new():
-    """
-    When a QbMultiUnits component is defined using only new/locally defined units, we can provide the
-    `propertyUrl` and the `valueUrl`.
-    """
-    column = QbColumn("Some Column", QbMultiUnits([NewQbUnit("Some New Unit")]))
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert (
-        "http://purl.org/linked-data/sdmx/2009/attribute#unitMeasure"
-        == default_property_uri
-    )
-    assert "cube-name.csv#unit/{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_multi_units_all_existing():
-    """
-    When a QbMultiUnits component is defined using just existing units, we can provide the `propertyUrl` and
-    `valueUrl`.
-    """
-    column = QbColumn(
-        "Some Column",
-        QbMultiUnits([ExistingQbUnit("http://base-uri/units/existing-unit")]),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert (
-        "http://purl.org/linked-data/sdmx/2009/attribute#unitMeasure"
-        == default_property_uri
-    )
-    assert "{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_multi_units_local_and_existing():
-    """
-    When a QbMultiUnits component is defined using a mixture of existing units and new units, we can't provide
-    an appropriate and consistent `valueUrl`.
-
-    An exception is raised when this is attempted.
-    """
-    column = QbColumn(
-        "Some Column",
-        QbMultiUnits(
-            [
-                NewQbUnit("Some New Unit"),
-                ExistingQbUnit("http://base-uri/units/existing-unit"),
-            ]
-        ),
-    )
-    with pytest.raises(Exception) as exception:
-
-        def fetch_exception():
-            empty_qbwriter._get_default_property_value_uris_for_column(column)
-
-        fetch_exception()
-
-
-def test_default_property_value_uris_multi_measure_all_new():
-    """
-    When a QbMultiMeasureDimension component is defined using only new/locally defined measures,
-    we can provide the `propertyUrl` and the `valueUrl`.
-    """
-    column = QbColumn(
-        "Some Column", QbMultiMeasureDimension([NewQbMeasure("Some New Measure")])
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "http://purl.org/linked-data/cube#measureType" == default_property_uri
-    assert "cube-name.csv#measure/{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_multi_measure_all_existing():
-    """
-    When a QbMultiUnits component is defined using just existing units, we can provide the `propertyUrl` and
-    `valueUrl`.
-    """
-    column = QbColumn(
-        "Some Column",
-        QbMultiMeasureDimension(
-            [ExistingQbMeasure("http://base-uri/measures/existing-measure")]
-        ),
-        csv_column_uri_template="http://base-uri/measures/{+some_column}",
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert "http://purl.org/linked-data/cube#measureType" == default_property_uri
-    assert "http://base-uri/measures/{+some_column}" == default_value_uri
-
-
-def test_default_property_value_uris_multi_measure_local_and_existing():
-    """
-    When a QbMultiUnits component is defined using a mixture of existing units and new units, we can't provide
-    an appropriate and consistent `valueUrl`.
-
-    An exception is raised when this is attempted.
-    """
-    column = QbColumn(
-        "Some Column",
-        QbMultiMeasureDimension(
-            [
-                NewQbMeasure("Some New Measure"),
-                ExistingQbMeasure("http://base-uri/measures/existing-measure"),
-            ]
-        ),
-    )
-
-    with pytest.raises(Exception) as exception:
-
-        def fetch_exception():
-            empty_qbwriter._get_default_property_value_uris_for_column(column)
-
-        fetch_exception()
-
-
-def test_default_property_value_uris_single_measure_obs_val():
-    """
-    There should be no `propertyUrl` or `valueUrl` for a `QbObservationValue`.
-    """
-    column = QbColumn(
-        "Some Column",
-        QbObservationValue(
-            unit=NewQbUnit("New Unit"), measure=NewQbMeasure("New Qb Measure")
-        ),
-    )
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = empty_qbwriter._get_default_property_value_uris_for_column(column)
-    assert default_property_uri == "cube-name.csv#measure/new-qb-measure"
-    assert default_value_uri is None
 
 
 def test_output_new_code_list_csvws_urls():
@@ -548,7 +74,7 @@ def test_output_new_code_list_csvws_urls():
         ) in graph
 
 
-def test_output_new_code_list_csvws_urls_with_uri_style_WithoutFileExtensions():
+def test_output_new_code_list_csvws_urls_with_uri_style_without_file_extensions():
     data = pd.DataFrame({"New Dimension": ["A", "B", "C"], "Value": [1, 2, 3]})
     cube = Cube(
         CatalogMetadata("Cube Name"),
@@ -574,57 +100,6 @@ def test_output_new_code_list_csvws_urls_with_uri_style_WithoutFileExtensions():
             URIRef("http://www.w3.org/ns/csvw#url"),
             Literal("some-dimension.csv", datatype=XSD.anyURI),
         ) in graph
-
-
-def test_default_property_value_uris_multi_measure_obs_val():
-    """
-    There should be no `valueUrl` for a `QbObservationValue`.
-    """
-    column = QbColumn("Some Column", QbObservationValue())
-
-    cube = Cube(
-        CatalogMetadata("Cube Name"),
-        pd.DataFrame({"Measure": ["kg"]}),
-        [QbColumn("Measure", QbMultiMeasureDimension([NewQbMeasure("kg")]))],
-    )
-    writer = QbWriter(cube)
-
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = writer._get_default_property_value_uris_for_column(column)
-    assert default_property_uri == "cube-name.csv#measure/{+measure}"
-    assert default_value_uri is None
-
-
-def test_default_property_value_uris_multi_existing_measure_obs_val():
-    """
-    The `propertyUrl` for a multi-existing-measure observation value should match the measure column's
-    value URI template.
-    """
-    column = QbColumn("Some Column", QbObservationValue())
-
-    cube = Cube(
-        CatalogMetadata("Cube Name"),
-        pd.DataFrame({"Measure": ["kg"]}),
-        [
-            QbColumn(
-                "Measure",
-                QbMultiMeasureDimension(
-                    [ExistingQbMeasure("http://some-existing-measures/kg")]
-                ),
-                csv_column_uri_template="http://some-existing-measures/{+measure}",
-            )
-        ],
-    )
-    writer = QbWriter(cube)
-
-    (
-        default_property_uri,
-        default_value_uri,
-    ) = writer._get_default_property_value_uris_for_column(column)
-    assert default_property_uri == "http://some-existing-measures/{+measure}"
-    assert default_value_uri is None
 
 
 def test_csv_col_definition_default_property_value_urls():
@@ -763,386 +238,6 @@ def test_csv_col_definition_suppressed():
     assert "valueUrl" not in csv_col
 
 
-def test_get_observation_value_col_for_title():
-    """
-    Ensure that the column title for the QbObservationValue returned matches the input column title.
-    """
-    expected_obs_val_col = QbColumn(
-        "Some Obs Val",
-        QbObservationValue(NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")),
-    )
-
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Some Dimension", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            expected_obs_val_col,
-        ],
-    )
-
-    writer = QbWriter(cube)
-    actual_obs_val_col = writer._get_observation_value_col_for_title("Some Obs Val")
-
-    assert actual_obs_val_col == expected_obs_val_col
-
-
-def test_get_observation_value_col_for_title_when_col_title_is_invalid():
-    """
-    Ensure that the invalid column title raises an exception.
-    """
-
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Invalid Col Title", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn(
-                "Some Obs Val",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-        ],
-    )
-
-    writer = QbWriter(cube)
-    with pytest.raises(Exception) as err:
-        writer._get_observation_value_col_for_title("Invalid Col Title")
-
-    assert (
-        str(err.value)
-        == 'Could not find one observation value column. Found 0 for title: "Invalid Col Title".'
-    )
-
-
-def test_get_cross_measures_slice_key_for_new_dimension():
-    """
-    Ensure that given a cube with NewQbDimension, the function returns the correct slice key.
-    """
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Some Dimension", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn(
-                "Some Obs Val",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-        ],
-    )
-    writer = QbWriter(cube)
-
-    actual_slice_key = writer._get_cross_measures_slice_key()
-    assert str(actual_slice_key.uri) == "cube.csv#slice/cross-measures"
-
-    component_properties = list(actual_slice_key.componentProperties)
-    assert len(component_properties) == 1
-    assert str(component_properties[0].uri) == "cube.csv#dimension/some-dimension"
-
-
-def test_get_cross_measures_slice_key_for_existing_dimension():
-    """
-    Ensure that given a cube with ExistingQbDimension, the function returns the correct slice key.
-    """
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn(
-                "Some Dimension",
-                ExistingQbDimension(
-                    "https://example.org/dimensions/existing_dimension"
-                ),
-            ),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn(
-                "Some Obs Val",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-        ],
-    )
-    writer = QbWriter(cube)
-
-    actual_slice_key = writer._get_cross_measures_slice_key()
-    assert str(actual_slice_key.uri) == "cube.csv#slice/cross-measures"
-
-    component_properties = list(actual_slice_key.componentProperties)
-    assert len(component_properties) == 1
-    assert (
-        str(component_properties[0].uri)
-        == "https://example.org/dimensions/existing_dimension"
-    )
-
-
-def test_is_cube_in_pivoted_shape_true_for_pivoted_shape_cube():
-    """
-    Ensure that the boolean returned for an input cube of pivoted shape is true.
-    """
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn(
-                "Some Dimension",
-                ExistingQbDimension(
-                    "https://example.org/dimensions/existing_dimension"
-                ),
-            ),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn(
-                "Some Obs Val",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-            QbColumn(
-                "Some Other Obs Val",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-        ],
-    )
-
-    writer = QbWriter(cube)
-    is_cube_pivoted = writer.is_cube_in_pivoted_shape
-    assert is_cube_pivoted == True
-
-
-def test_is_cube_in_pivoted_shape_false_for_standard_shape_cube():
-    """
-    Ensure that the boolean returned for an input cube of pivoted shape is false.
-    """
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn(
-                "Some Dimension",
-                ExistingQbDimension(
-                    "https://example.org/dimensions/existing_dimension"
-                ),
-            ),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn("Some Obs Val", QbObservationValue(unit=NewQbUnit("Some Unit"))),
-            QbColumn(
-                "Some Measure", QbMultiMeasureDimension([NewQbMeasure("New Measure")])
-            ),
-        ],
-    )
-
-    writer = QbWriter(cube)
-    is_cube_pivoted = writer.is_cube_in_pivoted_shape
-    assert is_cube_pivoted == False
-
-
-def test_is_cube_in_pivoted_shape_raise_exception():
-    """
-    Ensures that an exception is raised when the cube is in both the pivoted and standard shape.
-    """
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn(
-                "Some Dimension",
-                ExistingQbDimension(
-                    "https://example.org/dimensions/existing_dimension"
-                ),
-            ),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn("Some Obs Val", QbObservationValue(unit=NewQbUnit("Some Unit"))),
-            QbColumn(
-                "Some Other Obs Val",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-        ],
-    )
-
-    writer = QbWriter(cube)
-
-    with pytest.raises(Exception) as err:
-        writer.is_cube_in_pivoted_shape
-
-    assert str(err.value) == "The cube cannot be in both standard and pivoted shape"
-
-
-def test_get_observation_uri_for_pivoted_shape_data_set_new_qbmeasure():
-    """
-    Ensures that the observation value's URI is returned for an observation with a new Qbmeasure.
-    """
-    obs_val_column = QbColumn(
-        "Some Obs Val",
-        QbObservationValue(NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")),
-    )
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Some Dimension", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            obs_val_column,
-        ],
-    )
-    expected_observation_uri = "cube.csv#obs/{some_dimension}@some-measure"
-
-    writer = QbWriter(cube)
-    actual_observation_uri = writer._get_observation_uri_for_pivoted_shape_data_set(
-        obs_val_column
-    )
-
-    assert actual_observation_uri == expected_observation_uri
-
-
-def test_get_observation_uri_for_pivoted_shape_data_set_existing_qbmeasure():
-    """
-    Ensures that the observation value's URI is returned for an observation with a new Qbmeasure.
-    """
-    obs_val_column = QbColumn(
-        "Some Obs Val",
-        QbObservationValue(
-            ExistingQbMeasure("http://example.com/measures/existing_measure"),
-            NewQbUnit("Some Unit"),
-        ),
-    )
-
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Some Dimension", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            obs_val_column,
-        ],
-    )
-    expected_observation_uri = (
-        "cube.csv#obs/{some_dimension}@http-//example-com/measures/existing_measure"
-    )
-
-    writer = QbWriter(cube)
-    actual_observation_uri = writer._get_observation_uri_for_pivoted_shape_data_set(
-        obs_val_column
-    )
-
-    assert actual_observation_uri == expected_observation_uri
-
-
-def test_get_observation_uri_for_pivoted_shape_data_set_raise_exception():
-    """
-    Ensures that the observation value's URI raises an exception when given an unhandled QbMeasure type.
-    """
-
-    obs_val_column = QbColumn(
-        "Some Obs Val",
-        QbObservationValue(
-            TestQbMeasure("Some Measure"),
-            NewQbUnit("Some Unit"),
-        ),
-    )
-
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Some Dimension", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            obs_val_column,
-        ],
-    )
-    obs_val_measure = obs_val_column.structural_definition.measure
-
-    writer = QbWriter(cube)
-    with pytest.raises(Exception) as exception:
-        writer._get_observation_uri_for_pivoted_shape_data_set(obs_val_column)
-        assert (
-            str(exception.value) == f"Unhandled QbMeasure type {type(obs_val_measure)}"
-        )
-
-
-def test_get_pivoted_cube_slice_uri():
-    """
-    Ensures that the pivoted shape cube slice URI is returned.
-    """
-
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Some Dimension", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn(
-                "Some Obs Val",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"),
-                    NewQbUnit("Some Unit"),
-                ),
-            ),
-        ],
-    )
-
-    writer = QbWriter(cube)
-
-    expected_slice_uri = "cube.csv#slice/{some_dimension}"
-    actual_slice_uri = writer._get_pivoted_cube_slice_uri()
-
-    assert actual_slice_uri == expected_slice_uri
-
-
-def test_get_about_url_for_pivoted_shape_cube():
-    """
-    Ensures that the pivoted shape cube about URL is returned.
-    """
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Some Dimension", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn(
-                "Some Obs Val",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"),
-                    NewQbUnit("Some Unit"),
-                ),
-            ),
-        ],
-    )
-
-    writer = QbWriter(cube)
-    expected_about_url = "cube.csv#slice/{some_dimension}"
-    actual_about_url = writer._get_about_url()
-
-    assert actual_about_url == expected_about_url
-
-
-def test_get_about_url_for_standard_shape_cube():
-    """
-    Ensures that the standard shape cube about URL is returned.
-    """
-
-    cube = Cube(
-        CatalogMetadata("Cube"),
-        columns=[
-            QbColumn("Some Dimension", NewQbDimension("Some Dimension")),
-            QbColumn("Some Attribute", NewQbAttribute(label="Some Attribute")),
-            QbColumn(
-                "Some Obs Val",
-                QbObservationValue(
-                    unit=NewQbUnit("Some Unit"),
-                ),
-            ),
-            QbColumn(
-                "Some Measure", QbMultiMeasureDimension([NewQbMeasure("New Measure")])
-            ),
-        ],
-    )
-
-    writer = QbWriter(cube)
-
-    expected_about_url = "cube.csv#obs/{some_dimension}@{some_measure}"
-    actual_about_url = writer._uri.get_about_url()
-
-    assert actual_about_url == expected_about_url
-
-
 def test_virtual_columns_generated_for_single_obs_val():
     """
     Ensure that the virtual columns generated for a `QbObservationValue`'s unit and measure are correct.
@@ -1234,85 +329,6 @@ def test_virtual_columns_generated_for_multi_meas_obs_val():
         == virt_unit["propertyUrl"]
     )
     assert "cube-name.csv#unit/some-unit" == virt_unit["valueUrl"]
-
-
-def test_about_url_generation():
-    """
-    Ensuring that when an aboutUrl is defined for a non-multimeasure cube, the resulting URL
-    is built in the order in which dimensions appear in the cube.
-    """
-    data = pd.DataFrame(
-        {
-            "Existing Dimension": ["A", "B", "C"],
-            "Local Dimension": ["D", "E", "F"],
-            "Value": [2, 2, 2],
-        }
-    )
-
-    metadata = CatalogMetadata("Some Dataset")
-    columns = [
-        QbColumn(
-            "Existing Dimension",
-            ExistingQbDimension("https://example.org/dimensions/existing_dimension"),
-        ),
-        QbColumn(
-            "Local Dimension",
-            NewQbDimension.from_data("Name of New Dimension", data["Local Dimension"]),
-        ),
-        QbColumn(
-            "Value",
-            QbObservationValue(
-                ExistingQbMeasure("http://example.com/measures/existing_measure"),
-                NewQbUnit("New Unit"),
-            ),
-        ),
-    ]
-
-    cube = Cube(metadata, data, columns)
-
-    actual_about_url = QbWriter(cube)._get_about_url()
-    expected_about_url = "some-dataset.csv#slice/{existing_dimension},{local_dimension}"
-    assert actual_about_url == expected_about_url
-
-
-def test_about_url_generation_with_multiple_measures():
-    """
-    Ensuring that when an aboutUrl is defined for a multi-measure cube, the resulting URL
-    is built in the order in which dimensions appear in the cube except for the multi-measure
-    dimensions which are appended to the end of the URL.
-    """
-    data = pd.DataFrame(
-        {
-            "Measure": ["People", "Children", "Adults"],
-            "Existing Dimension": ["A", "B", "C"],
-            "Value": [2, 2, 2],
-            "Local Dimension": ["D", "E", "F"],
-            "Units": ["Percent", "People", "People"],
-        }
-    )
-
-    metadata = CatalogMetadata("Some Dataset")
-    columns = [
-        QbColumn(
-            "Measure", QbMultiMeasureDimension.new_measures_from_data(data["Measure"])
-        ),
-        QbColumn(
-            "Existing Dimension",
-            ExistingQbDimension("https://example.org/dimensions/existing_dimension"),
-        ),
-        QbColumn(
-            "Local Dimension",
-            NewQbDimension.from_data("Name of New Dimension", data["Local Dimension"]),
-        ),
-        QbColumn("Value", QbObservationValue("number")),
-        QbColumn("Units", QbMultiUnits.new_units_from_data(data["Units"])),
-    ]
-
-    cube = Cube(metadata, data, columns)
-
-    actual_about_url = QbWriter(cube)._get_about_url()
-    expected_about_url = "some-dataset.csv#slice/{existing_dimension},{local_dimension}"
-    assert actual_about_url == expected_about_url
 
 
 def test_serialise_new_attribute_values():
